Fix a compilation warning.

+++ sha1.h
@@ -308,8 +308,8 @@ SHA1_API void sha1_update(sha1_context* ctx, const void* src, size_t sz)
                 }
 
                 sha1_copy_memory(ctx->cache + ctx->cacheLen, bytes + totalBytesProcessed, bytesToProcess);
-                ctx->cacheLen       += bytesToProcess;
-                totalBytesProcessed += bytesToProcess;
+                ctx->cacheLen       += (unsigned int)bytesToProcess;    /* Safe cast. bytesToProcess will always be <= sizeof(ctx->cache) which is 64. */
+                totalBytesProcessed +=               bytesToProcess;
 
                 /* Update the number of bytes remaining in the cache so we can use it later. */
                 cacheRemaining = sizeof(ctx->cache) - ctx->cacheLen;
